Move build parameters out of CC impl package
