+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/internal/buildtree/DefaultBuildModelParameters.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 the original author or authors.
+ * Copyright 2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.internal.cc.impl.services
 
-import org.gradle.internal.buildtree.BuildModelParameters
-
+package org.gradle.internal.buildtree
 
 internal
 data class DefaultBuildModelParameters(
